feat!: Support float percentiles [#301]

+++ crates/hotpath-meta/src/lib_off.rs
@@ -240,7 +240,7 @@ impl HotpathGuardBuilder {
         Self {}
     }
 
-    pub fn percentiles(self, _percentiles: &[u8]) -> Self {
+    pub fn percentiles(self, _percentiles: &[f64]) -> Self {
         self
     }
